[fixed] more compiling problems
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/testing@1915 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/bazaar/BazaarManagerImplementation.cpp
@@ -535,7 +535,7 @@ AuctionItem* BazaarManagerImplementation::createAuctionItem(PlayerCreature* play
 	if (name.length() < 2)
 		objectName->getFullPath(name);
 
-	item->setLocation(planetStr, region, terminal->getObjectID(), terminal->getPositionX(), terminal->getPositionY(), false);
+	item->setLocation(planetStr, region, terminal->getObjectID(), (int)terminal->getPositionX(), (int)terminal->getPositionY(), false);
 	item->setOwnerID(player->getObjectID());
 	item->setOwnerName(playername);
 	item->setItemName(name);
